+++ Marshroute/Sources/Transitions/TransitionContexts/WeakTransitionsHandlerBox.swift
@@ -1,51 +1,48 @@
 public enum WeakTransitionsHandlerBox {
-    case animating(WeakBox<AnimatingTransitionsHandler>)
-    case containing(WeakBox<ContainingTransitionsHandler>)
+    case animating(WeakAnimatingTransitionsHandlerBox)
+    case containing(WeakContainingTransitionsHandlerBox)
     
     // MARK: - Init
     public init(transitionsHandlerBox: TransitionsHandlerBox)
     {
         switch transitionsHandlerBox {
-        case .animating(let strongBox):
-            let animatingTransitionsHandler = strongBox.unbox()
-            self = .init(animatingTransitionsHandler: animatingTransitionsHandler)
+        case .animating(let animatingTransitionsHandler):
+            self.init(animatingTransitionsHandler: animatingTransitionsHandler)
             
-        case .containing(let strongBox):
-            let containingTransitionsHandler = strongBox.unbox()
-            self = .init(containingTransitionsHandler: containingTransitionsHandler)
+        case .containing(let containingTransitionsHandler):
+            self.init(containingTransitionsHandler: containingTransitionsHandler)
         }
     }
     
     public init(animatingTransitionsHandler: AnimatingTransitionsHandler) {
-        self = .animating(WeakBox(animatingTransitionsHandler))
+        self = .animating(WeakAnimatingTransitionsHandlerBox(animatingTransitionsHandler))
     }
     
     public init(containingTransitionsHandler: ContainingTransitionsHandler) {
-        self = .containing(WeakBox(containingTransitionsHandler))
+        self = .containing(WeakContainingTransitionsHandlerBox(containingTransitionsHandler))
     }
     
     // MARK: - Public
     public func unbox() -> TransitionsHandler? {
         switch self {
-        case .animating(let animatingTransitionsHandlerBox):
-            return animatingTransitionsHandlerBox.unbox()
+        case .animating(let weakAnimatingTransitionsHandlerBox):
+            return weakAnimatingTransitionsHandlerBox.unbox()
             
-        case .containing(let containingTransitionsHandlerBox):
-            return containingTransitionsHandlerBox.unbox()
+        case .containing(let weakContainingTransitionsHandlerBox):
+            return weakContainingTransitionsHandlerBox.unbox()
         }
     }
-}
-
-public extension WeakTransitionsHandlerBox {
-    func transitionsHandlerBox() -> TransitionsHandlerBox? {
+    
+    // MARK: - Convenience
+    public func transitionsHandlerBox() -> TransitionsHandlerBox? {
         switch self {
-        case .animating(let animatingTransitionsHandlerBox):
-            if let animatingTransitionsHandler = animatingTransitionsHandlerBox.unbox() {
+        case .animating(let weakAnimatingTransitionsHandlerBox):
+            if let animatingTransitionsHandler = weakAnimatingTransitionsHandlerBox.unbox() {
                 return TransitionsHandlerBox(animatingTransitionsHandler: animatingTransitionsHandler)
             }
             
-        case .containing(let containingTransitionsHandlerBox):
-            if let containingTransitionsHandler = containingTransitionsHandlerBox.unbox() {
+        case .containing(let weakContainingTransitionsHandlerBox):
+            if let containingTransitionsHandler = weakContainingTransitionsHandlerBox.unbox() {
                 return TransitionsHandlerBox(containingTransitionsHandler: containingTransitionsHandler)
             }
         }
@@ -54,14 +51,32 @@ public extension WeakTransitionsHandlerBox {
     }
 }
 
-public extension TransitionsHandlerBox {
-    func weakTransitionsHandlerBox() -> WeakTransitionsHandlerBox {
-        switch self {
-        case .animating(let animatingTransitionsHandlerBox):
-            return WeakTransitionsHandlerBox(animatingTransitionsHandler: animatingTransitionsHandlerBox.unbox())
-            
-        case .containing(let containingTransitionsHandlerBox):
-            return WeakTransitionsHandlerBox(containingTransitionsHandler: containingTransitionsHandlerBox.unbox())
-        }
+// Эти врапперы существуют из-за ограничения Swift'а в виде ошибки `type A requires that type B be a class type`,
+// которая возникает в случаях, когда пытаешься сделать генерик тип `Box<T: AnimatingTransitionsHandler>` и в конструкторе
+// передаешь протокол вместо конкретного класса (не помогает даже сделать `init(t: some AnimatingTransitionsHandler)`
+
+public final class WeakAnimatingTransitionsHandlerBox {
+    private weak var animatingTransitionsHandler: AnimatingTransitionsHandler?
+
+    public init(_ animatingTransitionsHandler: AnimatingTransitionsHandler)
+    {
+        self.animatingTransitionsHandler = animatingTransitionsHandler
+    }
+    
+    public func unbox() -> AnimatingTransitionsHandler? {
+        animatingTransitionsHandler
+    }
+}
+
+public final class WeakContainingTransitionsHandlerBox {
+    private weak var containingTransitionsHandler: ContainingTransitionsHandler?
+
+    public init(_ containingTransitionsHandler: ContainingTransitionsHandler)
+    {
+        self.containingTransitionsHandler = containingTransitionsHandler
+    }
+    
+    public func unbox() -> ContainingTransitionsHandler? {
+        containingTransitionsHandler
     }
 }